refactor: use new page wrapper and adjust spacing
