+++ apps/journeys-admin/pages/reports/journeys.tsx
@@ -6,12 +6,11 @@ import {
   withAuthUserTokenSSR
 } from 'next-firebase-auth'
 import { NextSeo } from 'next-seo'
-import { useRouter } from 'next/router'
 import { serverSideTranslations } from 'next-i18next/serverSideTranslations'
 import { useTranslation } from 'react-i18next'
 import { getLaunchDarklyClient } from '@core/shared/ui/getLaunchDarklyClient'
 import Box from '@mui/material/Box'
-import { PageWrapper } from '../../src/components/PageWrapper'
+import { PageWrapper } from '../../src/components/NewPageWrapper'
 import i18nConfig from '../../next-i18next.config'
 import { MemoizedDynamicReport } from '../../src/components/DynamicPowerBiReport'
 import { JourneysReportType } from '../../__generated__/globalTypes'
@@ -19,7 +18,6 @@ import { useTermsRedirect } from '../../src/libs/useTermsRedirect/useTermsRedire
 import { ReportsNavigation } from '../../src/components/ReportsNavigation'
 
 function ReportsJourneysPage(): ReactElement {
-  const router = useRouter()
   const { t } = useTranslation('apps-journeys-admin')
   const AuthUser = useAuthUser()
 
@@ -28,11 +26,7 @@ function ReportsJourneysPage(): ReactElement {
   return (
     <>
       <NextSeo title={t('Journeys Report')} />
-      <PageWrapper
-        title={t('Journeys Report')}
-        authUser={AuthUser}
-        router={router}
-      >
+      <PageWrapper title={t('Journeys Report')} authUser={AuthUser}>
         <Box sx={{ height: 'calc(100vh - 48px)' }}>
           <ReportsNavigation selected="journeys" />
           <MemoizedDynamicReport reportType={JourneysReportType.multipleFull} />
